requirements.txt
imageio==2.34.2
imageio-ffmpeg==0.5.1
matplotlib==3.9.1
pillow==10.4.00
torch==2.3.1
torchvision==0.18.1